+++ shows/nfo.go
@@ -1,99 +1,10 @@
 package shows
 
 import (
-	"encoding/xml"
 	"fengqi/kodi-metadata-tmdb-cli/tmdb"
 	"fengqi/kodi-metadata-tmdb-cli/utils"
 	"strconv"
 )
-
-// TvShowNfo tvshow.nfo
-//
-// https://kodi.wiki/view/NFO_files/TV_shows
-// NFO files for TV Shows are a little bit more complex as they require the following NFO files:
-//
-// One nfo file for the TV Show. This file holds the overall TV show information
-// One nfo file for each Episode. This file holds information specific to that episode
-// For one TV Show with 10 episodes, 11 nfo files are required.
-type TvShowNfo struct {
-	XMLName        xml.Name      `xml:"tvshow"`
-	Title          string        `xml:"title"`
-	OriginalTitle  string        `xml:"originaltitle"`
-	ShowTitle      string        `xml:"showtitle"` // 	Not in common use, but some skins may display an alternate title
-	SortTitle      string        `xml:"sorttitle"`
-	Ratings        Ratings       `xml:"ratings"`
-	UserRating     float32       `xml:"userrating"`
-	Top250         string        `xml:"-"`
-	Season         int           `xml:"season"`
-	Episode        int           `xml:"episode"`
-	DisplayEpisode int           `xml:"-"`
-	DisplaySeason  int           `xml:"-"`
-	Outline        string        `xml:"-"`
-	Plot           string        `xml:"plot"`
-	Tagline        string        `xml:"-"`
-	Runtime        int           `xml:"-"`
-	Thumb          []Thumb       `xml:"-"`
-	FanArt         FanArt        `xml:"fanart"`
-	MPaa           string        `xml:"mpaa"`
-	PlayCount      int           `xml:"-"`
-	LastPlayed     string        `xml:"-"`
-	EpisodeGuide   EpisodeGuide  `xml:"-"`
-	Id             int           `xml:"id"`
-	UniqueId       UniqueId      `xml:"uniqueid"`
-	Genre          []string      `xml:"genre"`
-	Tag            []string      `xml:"tag"`
-	Premiered      string        `xml:"premiered"`
-	Year           string        `xml:"-"`
-	Status         string        `xml:"status"`
-	Aired          string        `xml:"-"`
-	Studio         []string      `xml:"studio"`
-	Trailer        string        `xml:"trailer"`
-	Actor          []Actor       `xml:"actor"`
-	NamedSeason    []NamedSeason `xml:"namedseason"`
-	Resume         Resume        `xml:"-"`
-	DateAdded      int           `xml:"-"`
-}
-
-type UniqueId struct {
-	XMLName xml.Name `xml:"uniqueid"`
-	Type    string   `xml:"type,attr"`
-	Default bool     `xml:"default,attr"`
-}
-
-type Actor struct {
-	Name      string `xml:"name"`
-	Role      string `xml:"role"`
-	Order     int    `xml:"order"`
-	SortOrder int    `xml:"sortorder"`
-	Thumb     string `xml:"thumb"`
-}
-
-type FanArt struct {
-	XMLName xml.Name    `xml:"fanart"`
-	Thumb   []ShowThumb `xml:"thumb"`
-}
-
-type ShowThumb struct {
-	Preview string `xml:"preview,attr"`
-}
-
-type EpisodeGuide struct {
-	Url Url `xml:"url"`
-}
-
-type Url struct {
-	Cache string `xml:"cache"`
-}
-
-type NamedSeason struct {
-	Number int    `xml:"number,attr"`
-	Value  string `xml:",chardata"`
-}
-
-type Resume struct {
-	Position string `xml:"position"`
-	Total    int    `xml:"total"`
-}
 
 func (d *Dir) saveToNfo(detail *tmdb.TvDetail) error {
 	utils.Logger.InfoF("save tvshow.nfo to: %s", d.getNfoFile())
@@ -168,4 +79,60 @@ func (d *Dir) saveToNfo(detail *tmdb.TvDetail) error {
 	}
 
 	return utils.SaveNfo(d.getNfoFile(), top)
+}
+
+// SaveTvEpisodeNFO 保存每集的信息到独立的NFO文件
+func (f *File) saveToNfo(episode *tmdb.TvEpisodeDetail) error {
+	utils.Logger.InfoF("save episode nfo to: %s", f.getNfoFile())
+
+	actor := make([]Actor, 0)
+	for _, item := range episode.GuestStars {
+		actor = append(actor, Actor{
+			Name:      item.Name,
+			Role:      item.Character,
+			Order:     item.Order,
+			Thumb:     tmdb.ImageW500 + item.ProfilePath,
+			SortOrder: item.Order,
+		})
+	}
+
+	// 评分
+	rating := make([]Rating, 1)
+	rating[0] = Rating{
+		Name:  "tmdb",
+		Max:   10,
+		Value: episode.VoteAverage,
+		Votes: episode.VoteCount,
+	}
+
+	top := &TvEpisodeNfo{
+		Title:         episode.Name,
+		ShowTitle:     episode.Name,
+		OriginalTitle: episode.Name,
+		Plot:          episode.Overview,
+		UniqueId: UniqueId{
+			Type:    strconv.Itoa(episode.Id),
+			Default: true,
+		},
+		Premiered:      episode.AirDate,
+		MPaa:           "TV-14",
+		Season:         episode.SeasonNumber,
+		Episode:        episode.EpisodeNumber,
+		DisplaySeason:  episode.SeasonNumber,
+		DisplayEpisode: episode.EpisodeNumber,
+		UserRating:     episode.VoteAverage,
+		//Tagline:        "111",
+		TmdbId:  "tmdm" + strconv.Itoa(episode.Id),
+		Runtime: 6,
+		Status:  "ok",
+		Actor:   actor,
+		Thumb: Thumb{
+			Aspect:  "thumb",
+			Preview: tmdb.ImageOriginal + episode.StillPath,
+		},
+		Ratings: rating,
+		Aired:   episode.AirDate,
+	}
+
+	return utils.SaveNfo(f.getNfoFile(), top)
 }
